purchase hung

+++ GiftShop/src/java/models/ProductFacade.java
@@ -108,8 +108,8 @@ public class ProductFacade extends AbstractFacade<Product> implements ProductFac
     }  
     
     @Override
-    public int getMaxPrice() {
-        return (int) em.createQuery("select max(p.proPrice) from Product p").getSingleResult();
+    public Double getMaxPrice() {
+        return (Double) em.createQuery("select max(p.proPrice) from Product p").getSingleResult();
     }
 
     @Override
